+++ cloudinit/sources/DataSourceEc2.py
@@ -13,13 +13,13 @@ import os
 import time
 
 from cloudinit import dmi
-from cloudinit import ec2_utils as ec2
 from cloudinit import log as logging
 from cloudinit import net, sources
 from cloudinit import url_helper as uhelp
 from cloudinit import util, warnings
 from cloudinit.event import EventScope, EventType
 from cloudinit.net.dhcp import EphemeralDHCPv4, NoDHCPLeaseError
+from cloudinit.sources.helpers import ec2
 
 LOG = logging.getLogger(__name__)
