+++ cloudinit/sources/helpers/cloudsigma.py
