+++ cloudinit/sources/helpers/ec2.py
